requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.1.4
django-debug-toolbar==4.4.6
django-filter==24.3
djangorestframework==3.15.2
Markdown==3.7
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
